\name{MPCMR_GMM}
\alias{MPCMR_GMM}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Individual-level data time-varying MR
}
\description{
Multiple-principal-component Mendelian randomization fitting with GMM methods desgined for individual-level data. This function supports overlapping-data setting, where the overlapping data will first clumped to one-sample and continues the fitting afterward.
}
\usage{
MPCMR_GMM( Gmatrix,
           res,
           Yvector,
           Gymatrix=NA,
           IDmatch=NA,
           nPC=NA,
           nL=NA,
           eigenfit=TRUE,
           polyfit=TRUE,
           LMCI=TRUE,
           LMCI2=TRUE,
           nLM=20,
           Parallel=TRUE,
           cores_used=NA,
           XYmodel=NA
         )
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Gmatrix}{
a matrix indicates the instrument information
}
\item{res}{
a list result contains the principal component matrix and the eigenfunctions. \code{res} is usually derived by \code{FPCA} from the package fdapace. \code{res} must contains the following elements:
\code{res$cumFVE}: the vector correpsonding the cumulative fraction-of-variance-explained (FVE).
\code{res$xiEst}: the matrix contains the principal components values for each individual, where the rows correspond to the individuals and columns correspond to the peincipal compoemnts.
}
\item{Yvector}{
a vector indicates the individual point outcome
}
\item{Gymatrix}{
a matrix indicates the instrument information for the outcome data. The default value is \code{NA}, which corresponds to the one-sample setting.
}
\item{IDmatch}{
a vector indicates the overlapping data index. This is better generated by \code{match()}. The default value is \code{NA}, correponding to the one-sample setting. If you have an overlapping samples, and the ID vector for the exposure and outcome data are \code{ID_X} and \code{ID_Y}, you can get the  \code{IDmatch} as \code{myIDmatch <- match( ID_X, ID_Y)}.
}
\item{nPC}{
a integer indicates the number of principal components used for MPCMR fitting. The default value is the number of principal compoments that just explain more than 95 percent variations.
}
\item{nL}{
a integer indicates the degree of the polynomial (the number of polynomial basisfunctions)
}
\item{eigenfit}{
logic. Whether to do MPCMR fitting with the eigenfunction as the basis function.
}
\item{polyfit}{
logic. Whether to do MPCMR fitting with the polynomial set as the basis function.
}
\item{LMCI}{
logic. Whether to calculate the CI with LM statistic where the basisfunction is the eigenfunction.
}
\item{LMCI2}{
logic. Whether to calculate the CI with LM statistic where the basisfunction is the polynomial.
}
\item{nLM}{
a integer indicates the number of increasing point for each dimension when calculating the CI with LM statistic.
}
\item{Parallel}{
logic. Whether to use parallel computing. The default value is \code{TRUE}.
}
\item{cores_used}{
an positive integer indicates how much cores will be used for the parallel computing. Only work when \code{Parallel=TRUE}. The default core number is the maximal cores number minus one.
}
\item{XYmodel}{
a character indicates the XY model. It should be only used for simulation purpose.
}

}
\details{
Note that you should have individual-level data containing the genetic variants (i.e. genotype) information and the longitudinal information of the exposure of interest. Both information should be contained simultaneously for each individual.

The longitudinal information must contain both the exposure level and its corresponding measured time point (age). It allows for the exposure to be measured at different time points (ages) for every individual, and each individual can have a sparse measurement.

You will also have the outcome data, which can be summary-data or individual-data; one-sample or two-sample as the exposure data. Depending on the specific data setting, you can use different functions of TVMR. If you outcome data is individual-level data that is one-sample or overlapping sample with your exposure data, you just need to use \code{MPCMR_GMM}. If your outcome data is summary information or from two sample as the exposure data, you will need to use \code{MPCMR_GMM_twosample}.


}
\value{
\code{MPCMR_GMM} retuns a list, consisting of various results, inclduing fitted parameters (and their standard errors), the weak IV assessment results, the IV validity assessment results, the fitted curve.

\item{nPC_used}{how many principal components were used in MPCMR.}
\item{L}{the number of polynomial basis function used for fitting MPCMR.}
\item{K}{the number of eigenfunction basis function used for fitting MPCMR.}
\item{ISres}{the table results of the instrument strength. The columns are coefficient of determination, F value, conditional F, Q statistic value, degree-of-freedom of Q, and the p-value, respectively.}
\item{scatterp}{the MR scatter plot corresponding to the genetic association with the first and second principal compoments.}
\item{one_sample_size}{the sample size of the data finally used for MPCMR fitting.}
\item{IV_validity_test}{the IV validity test results, where the three values are the Q statistic, the degree-of-freedom and the p-value, respectively.}
\item{MPCMRest}{the fitted parameters for the eigenfunction basis set.}
\item{MPCMRvar}{the corresponding variance matrix of \code{MPCMRest}.}
\item{p1}{the fitted curve with eigenfunctin as the bsis function.}
\item{ggdata1}{the dataframe used for producing \code{p1} via ggplot.}
\item{p2}{the fitted curve with polynomial as the bsis function.}
\item{ggdata2}{the dataframe used for producing \code{p2} via ggplot.}
\item{IV_validity_and_basisfunction_test}{the IV validity test results considering the parameteric (polynomial) basis function,  where the three values are the Q statistic, the degree-of-freedom and the p-value, respectively.}

\code{SE},\code{MSE},\code{Co},\code{Coverage_rate},\code{Co_LM},\code{Coverage_rate_LM},\code{sig_points},\code{sig_points_LM} gives some useful information when the true effect function is known (given by the argument \code{XYmodel}). They are used for simulation design.

The result name with the complementary symbol \code{_p} represents the results when the basis function are polynomial functions. For example, \code{MPCMRest_p} is the the fitted parameters for the polynmomial functions.
}

\author{
Haodong Tian
}

\examples{
###see README.md file and TVMR/sim_real_illustration/MPCMR_illustration.R from GitHub
}

